requirements.txt
python-dotenv==1.0.0
streamlit==1.18.1
openai 
pytube3==9.6.4
yt-dlp